requirements.txt
pyserial
python-jwt>=3.3.4
requests>=2.20.0
Pyrebase
pickle5
Pillow
RPi.GPIO
opencv-python
streamlit==0.62.0
orjson
maturin